refactor: consolidate processors tests with v5/v6 model support
